Updating client facing errors
