+++ crates/synthesis/src/permutation/utils.rs
@@ -10,12 +10,32 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
-use crate::permutation::PermutationError;
+use crate::PyErr;
 use ndarray::ArrayViewMut1;
 use ndarray::{Array1, ArrayView1};
+use pyo3::exceptions::PyValueError;
 use std::vec::Vec;
 
 use qiskit_util::py::PySequenceIndex;
+
+/// Possible errors that can occur when validating a permutation pattern.
+#[derive(Debug, thiserror::Error)]
+pub(crate) enum PermutationError {
+    #[error("invalid permutation: input contains a negative number")]
+    NegativeEntry,
+
+    #[error("invalid permutation: input has length {length} and contains {value}")]
+    OutOfBounds { length: usize, value: i64 },
+
+    #[error("invalid permutation: input contains {value} more than once")]
+    DuplicateEntry { value: i64 },
+}
+
+impl From<PermutationError> for PyErr {
+    fn from(value: PermutationError) -> Self {
+        PyValueError::new_err(value.to_string())
+    }
+}
 
 pub fn validate_permutation(pattern: &ArrayView1<i64>) -> Result<(), PermutationError> {
     let n = pattern.len();
